Enhanced User Experience: Register Options, Admin Panel, and Default Roles
In this update, I made some big improvements to how the app looks and handles user information. Now, when people sign up, they can choose if they're a student or a teacher, and I added a special login for administrators. The main page and the way everything is set up now look nicer and work better. I created a special section just for administrators to manage important tasks. I also set up default roles like "Student," "Teacher", "Moderator" and "Admin," and made a default admin account with the right permissions. I tested signing up and logging in with different types of users to make sure everything works smoothly. This update is a big step forward in making the app more organized and user-friendly. I will keep working on it to make sure it's the best it can be!