Resolved Admin Interface Bug and Designed Student Area for this project.
In this update, I focused on making our application safer. I  fixed a bug that was causing issues with how admins log in, ensuring a stronger and more reliable admin login process. I also made good progress in building the student section. I added a special "Question Controller" to help students interact with the system, and I created a common layout to make everything look the same and easy to use. These changes aim to make this app more secure and user-friendly.